change from Double to Long (price is in cents here) fix products total price sum (wasn't accounting for product quantities)

+++ src/main/java/com/reactive/order/Product.java
@@ -15,7 +15,7 @@ public class Product {
     @Indexed
     private Long id;
     private String name;
-    private Double price;
+    private Long price;
     private Long quantity;
 
 }
